Split piptols-python cli: "cli.py" and "dev-cli.py"
e.g.: Installing packages on a Raspberry Pi takes some times and binary packages can make troubles.
See: https://github.com/jedie/inverter-connect/issues/22

So split between "normal" and "dev" mode is a good idea ;)
